Cottage owner authorization all done

+++ src/main/java/stasaaleksadavid/isabackend/controller/RoomController.java
@@ -25,49 +25,62 @@ public class RoomController {
     public List<Room> getAllRooms(){return roomRepository.findAll();}
 
     //create
-    @PostMapping("/rooms")
-    public  Room createRoom(@RequestBody Room room){
-        return roomRepository.save(room);
+    @PostMapping("/rooms/{type}")
+    public  Room createRoom(@PathVariable String type,@RequestBody Room room){
+        if ( type.equals("admin") || type.equals("cottage_owner") || type.equals("main_admin")) {
+            return roomRepository.save(room);
+        }
+        else return null;
     }
 
     //get by id
-    @GetMapping("/rooms/{id}")
-    public ResponseEntity <Room> getRoomById(@PathVariable Long id){
-        Room room = roomRepository.findById(id).orElseThrow(() -> new ResourceNotFoundException("Room does not exist with id:"+ id));
-        return ResponseEntity.ok(room);
+    @GetMapping("/rooms/{type}/{id}")
+    public ResponseEntity <Room> getRoomById(@PathVariable String type,@PathVariable Long id){
+        if ( type.equals("admin") || type.equals("cottage_owner") || type.equals("main_admin")) {
+            Room room = roomRepository.findById(id).orElseThrow(() -> new ResourceNotFoundException("Room does not exist with id:" + id));
+            return ResponseEntity.ok(room);
+        }
+        else return null;
     }
 
     //update
-    @PutMapping("/rooms/{id}")
-    public ResponseEntity<Room> updateRoom(@PathVariable Long id,@RequestBody Room roomDetails){
-        Room room = roomRepository.findById(id).orElseThrow(() -> new ResourceNotFoundException("Room does not exist with id:"+ id));
+    @PutMapping("/rooms/{type}/{id}")
+    public ResponseEntity<Room> updateRoom(@PathVariable String type,@PathVariable Long id,@RequestBody Room roomDetails){
+        if ( type.equals("admin") || type.equals("cottage_owner") || type.equals("main_admin")) {
+            Room room = roomRepository.findById(id).orElseThrow(() -> new ResourceNotFoundException("Room does not exist with id:" + id));
 
 
-        room.setNumber(roomDetails.getNumber());
-        room.setNumberOfBeds(roomDetails.getNumberOfBeds());
-        room.setDescription(roomDetails.getDescription());
+            room.setNumber(roomDetails.getNumber());
+            room.setNumberOfBeds(roomDetails.getNumberOfBeds());
+            room.setDescription(roomDetails.getDescription());
 
-        Room updatedRoom = roomRepository.save(room);
-        return ResponseEntity.ok(updatedRoom);
+            Room updatedRoom = roomRepository.save(room);
+            return ResponseEntity.ok(updatedRoom);
+        }
+        else return null;
     }
 
 
     //delete
-    @DeleteMapping("/rooms/{id}")
-    public Map<String, Boolean> deleteRoom(@PathVariable Long id){
+    @DeleteMapping("/rooms/{type}/{id}")
+    public Map<String, Boolean> deleteRoom(@PathVariable Long id,@PathVariable String type) {
+        if (type.equals("ship_owner") || type.equals("admin") || type.equals("cottage_owner") || type.equals("fishing_instructor") || type.equals("main_admin")) {
+            Room room = roomRepository.findById(id).orElseThrow(() -> new ResourceNotFoundException("Room does not exist with id:" + id));
 
-        Room room = roomRepository.findById(id).orElseThrow(() -> new ResourceNotFoundException("Room does not exist with id:"+ id));
-
-        roomRepository.delete(room);
-        Map<String, Boolean> response = new HashMap<>();
-        response.put("deleted", Boolean.TRUE);
-        return (Map<String, Boolean>) ResponseEntity.ok(response);
+            roomRepository.delete(room);
+            Map<String, Boolean> response = new HashMap<>();
+            response.put("deleted", Boolean.TRUE);
+            return (Map<String, Boolean>) ResponseEntity.ok(response);
+        }
+        else return null;
     }
 
-
-
-    @GetMapping("/rooms/cottage/{cottageid}")
-    public List<Room> getRoomByCottageId(@PathVariable Long cottageid){
+//get by cottage id
+    @GetMapping("/rooms/cottage/{type}/{cottageid}")
+    public List<Room> getRoomByCottageId(@PathVariable String type,@PathVariable Long cottageid){
+        if(type.equals("ship_owner") || type.equals("admin")|| type.equals("cottage_owner")|| type.equals("fishing_instructor")|| type.equals("main_admin")){
         return roomRepository.findByCottageId(cottageid);
+    }
+        else return null;
     }
 }
